reorganising the project folders and adding an autoloading class to include necessary files dynamicaly
